+++ lib/util/constants.dart
@@ -1,0 +1,12 @@
+const flavorDefault = 'all';
+const supportedFolders = [
+  'colors',
+  'configs',
+  'fonts',
+  'images',
+  'raws',
+  'scripts',
+  'translations',
+  'values',
+  'vectors',
+];