+++ lib/util/filename_util.dart
@@ -1,0 +1,12 @@
+import 'package:assets_generator/util/dart_keywords.dart';
+
+String safeName(String name) {
+  var trimmed = name
+      .trim()
+      .replaceAll(RegExp(r'\W'), '_')
+      .replaceAll(RegExp('_{2,}'), '_');
+  if (dartKeywords.contains(trimmed)) {
+    trimmed = '${trimmed}_';
+  }
+  return trimmed;
+}