+++ CHANGELOG.md
@@ -23,6 +23,23 @@
 
 - Entries now sorted by keys.
 
-# 1.4.0
+## 1.4.0
 
 - Add support for multi-leveled (map and list) config and value resources.
+
+## 1.4.1
+
+- Add `raws/` folder.
+
+## 2.0.0
+
+- Add support for flavored assets.
+- Generator now also update pubspec.yaml according to assets folder.
+
+## 2.0.1
+
+- Avoid generating fields for hidden files.
+
+## 2.0.2
+
+- Add more image formats.